TAMU Migrate: Update Dockerfile to contain needed dependencies.
Ruby is needed with the postgresql gem to perform the migration from the within a running Docker instance.
Java and related packaging tools are needed.
Several useful tools are added to assist when operating on the Docker instance.

Don't bother with running tests when building.
Make sure required files needed to be run have the executable bit set.